Restructure OOP to fix import loop issue
