--- snn_voice/model/mx/__init__.py
@@ -1,17 +1,0 @@
-from .m5_cnn import M5CNN
-from .m5_snn import m5_snn_init
-from .m5_snn_latency import M5SNNLatency
-from .m5_snn_rate import M5SNNRate
-from .m5_snn_repeat import M5SNNRepeat
-from .mx_cnn_block import MxCNNBlock
-from .mx_snn_block import MxSNNBlock
-
-__all__ = [
-    'M5CNN',
-    'm5_snn_init',
-    'M5SNNRate',
-    'M5SNNRepeat',
-    'M5SNNLatency',
-    'MxCNNBlock',
-    'MxSNNBlock',
-]
