+++ snn_voice/model/mx/m5/m5_snn_latency.py
@@ -1,5 +1,5 @@
 from snn_voice.model.module import ModuleSNNLatency
-from snn_voice.model.mx import m5_snn_init
+from snn_voice.model.mx.m5 import m5_snn_init
 
 
 class M5SNNLatency(ModuleSNNLatency):
