+++ snn_voice/model/piczak/__init__.py
@@ -1,7 +1,5 @@
 from .piczak_cnn import PiczakCNN
-from .piczak_cnn_block import PiczakCNNBlock
 from .piczak_snn import piczak_snn_init
-from .piczak_snn_block import PiczakSNNBlock
 from .piczak_snn_latency import PiczakSNNLatency
 from .piczak_snn_rate import PiczakSNNRate
 from .piczak_snn_repeat import PiczakSNNRepeat
@@ -12,6 +10,4 @@ __all__ = [
     'PiczakSNNLatency',
     'PiczakSNNRate',
     'PiczakSNNRepeat',
-    'PiczakCNNBlock',
-    'PiczakSNNBlock',
 ]
